backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dateutil
firebase-admin
python-dotenv
